After Release 1.3.2 = Change State from DEV to PUBLIC = Fixed Converting Error

+++ src/main/java/de/presti/ree6/commands/impl/level/Leaderboards.java
@@ -64,7 +64,7 @@ public class Leaderboards extends Command {
         return level;
     }
 
-    public int getLevelVC(int xp) {
+    public int getLevelVC(long xp) {
         int level = 1;
 
         while (xp > 1000) {
